requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==23.12.1
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
dictdiffer==0.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.8
iso8601==1.1.0
isort==5.13.2
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pycparser==2.22
pydantic==2.9.0b1
pydantic-settings==2.4.0
pydantic_core==2.23.0
Pygments==2.18.0
PyJWT==2.9.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rich==13.8.0
ruff==0.0.281
setuptools==70.3.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tomlkit==0.13.2
tortoise-orm==0.20.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.6
# uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0


python-magic==0.4.27
aiofiles==23.2.1
Pillow==10.0.0
opencv-python== 4.8.0.74